# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venom/exports/{kind}:
    get:
      operationId: exportVenomPersonalMarkdown
      tags: [venom]
      summary: Download your personal Brain notes or SOPs as Markdown
      description: >-
        Always scoped to the requesting account's personal tier; contains no
        workspace content, so it keeps working after leaving a workspace.
      parameters:
        - $ref: "#/components/parameters/VenomExportKind"
        - name: scope
          in: query
          required: false
          description: >-
            Brain exports only. `sorted` limits the download to knowledge
            outside the Unsorted holding area; `unsorted` downloads just the
            Unsorted items. Omit for everything.
          schema:
            type: string
            enum: [sorted, unsorted]
      responses:
        "200":
          description: Markdown attachment
          content:
            text/markdown:
              schema: { type: string, maxLength: 400000 }
        "401": { description: Authentication required }
...
  /venom/apps/{appId}/ai:
    get:
      operationId: getVenomAppAi
      tags: [venom]
      summary: Owner view of an app's whitelabeled AI usage and controls
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: Month-to-date usage, spend controls, and credential summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomAppAiOverview"
        "401":
          description: Authentication required
        "404":
          description: App not found
  /venom/apps/{appId}/ai/settings:
    put:
      operationId: updateVenomAppAiSettings
      tags: [venom]
      summary: Set the app's monthly AI spending cap and pause switch
      parameters:
        - $ref: "#/components/parameters/AppId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenomAppAiSettingsUpdate"
      responses:
        "200":
          description: Updated AI overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomAppAiOverview"
        "400":
          description: Invalid settings update
        "401":
          description: Authentication required
        "404":
          description: App not found
  /venom/apps/{appId}/ai/credential/rotate:
    post:
      operationId: rotateVenomAppAiCredential
      tags: [venom]
      summary: Rotate the app's AI gateway credential
      description: >-
        Revokes the current credential and mints a replacement in one step.
        The new secret is delivered into the provisioned app's secret storage
        through the provisioning provider — it never appears in this API. If
        immediate delivery is not possible, the credential is delivered at
        the next provisioning handoff and the overview reports it as not yet
        delivered.
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: AI overview reflecting the fresh credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomAppAiOverview"
        "401":
          description: Authentication required
        "404":
          description: App not found
  /venom/apps/{appId}/ai/credential/revoke:
    post:
      operationId: revokeVenomAppAiCredential
      tags: [venom]
      summary: Revoke the app's AI gateway credential
      description: >-
        Server-side kill switch: the deployed app's AI stops immediately and
        stays off until a rotate or the next provisioning handoff mints a
        replacement.
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: AI overview with no active credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenomAppAiOverview"
        "401":
          description: Authentication required
        "404":
          description: App not found
...
components:
  parameters:
...
    AppId:
      name: appId
      in: path
      required: true
      schema:
        type: string
        pattern: "^[0-9a-fA-F-]{36}$"
...
    VenomExportKind:
      name: kind
      in: path
      required: true
      schema:
        type: string
        enum: [brain, sops]
...
  schemas:
...
    Uuid:
      type: string
      pattern: "^[0-9a-fA-F-]{36}$"
...
    VenomAppAiOverview:
      type: object
      additionalProperties: false
      description: >-
        Owner-only view of an app's whitelabeled AI: month-to-date usage from
        the canonical ledger, spend controls, and a summary of the gateway
        credential. Never contains the credential secret, provider names, or
        per-token rates — costs are aggregated dollars and models are Venom
        aliases only.
      required:
        - appId
        - paused
        - monthlyCapUsd
        - safetyCapUsd
        - credential
        - usage
        - ownerMonthUsd
      properties:
        appId: { $ref: "#/components/schemas/Uuid" }
        paused:
          type: boolean
        monthlyCapUsd:
          oneOf:
            - type: number
              minimum: 0
            - type: "null"
        safetyCapUsd:
          type: number
          minimum: 0
        credential:
          oneOf:
            - $ref: "#/components/schemas/VenomAppAiCredentialSummary"
            - type: "null"
        usage:
          $ref: "#/components/schemas/VenomAppAiUsage"
        ownerMonthUsd:
          type: number
          minimum: 0
    VenomAppAiCredentialSummary:
      type: object
      additionalProperties: false
      description: >-
        Display-safe credential summary. The prefix identifies the key for
        the owner; the secret itself is only ever delivered into the
        provisioned app's secret storage.
      required: [displayPrefix, createdAt, lastUsedAt, delivered]
      properties:
        displayPrefix:
          type: string
          maxLength: 20
        createdAt:
          type: string
          format: date-time
        lastUsedAt:
          type: [string, "null"]
          format: date-time
        delivered:
          type: boolean
    VenomAppAiUsage:
      type: object
      additionalProperties: false
      required:
        - periodStart
        - periodEnd
        - costUsd
        - requests
        - promptTokens
        - outputTokens
        - hasEstimates
        - models
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        costUsd:
          type: number
          minimum: 0
        requests:
          type: number
          multipleOf: 1
          minimum: 0
        promptTokens:
          type: number
          multipleOf: 1
          minimum: 0
        outputTokens:
          type: number
          multipleOf: 1
          minimum: 0
        hasEstimates:
          type: boolean
        models:
          type: array
          maxItems: 40
          items:
            $ref: "#/components/schemas/VenomAppAiModelUsage"
    VenomAppAiModelUsage:
      type: object
      additionalProperties: false
      required: [modelId, modelName, costUsd, requests]
      properties:
        modelId:
          type: string
          maxLength: 80
        modelName:
          type: string
          maxLength: 120
        costUsd:
          type: number
          minimum: 0
        requests:
          type: number
          multipleOf: 1
          minimum: 0
    VenomAppAiSettingsUpdate:
      type: object
      additionalProperties: false
      description: >-
        Full settings write: both fields are explicit on every update. The
        cap is dollars per calendar month (UTC); null removes the owner cap
        while the global safety cap keeps applying.
      required: [monthlyCapUsd, paused]
      properties:
        monthlyCapUsd:
          oneOf:
            - type: number
              minimum: 0.01
              maximum: 100000
            - type: "null"
        paused:
          type: boolean
...